Add code to environment to build under RISC OS.
Each of the components has a small obey file to build a RISC OS
structured source tree. A Makefile will build the libraries or the
executable in each directory.

The cJSON test code builds and runs, but then aborts with a division
by zero error, because it intentionally divides by zero - assuming it
will get back Infinity. Meh.

The libwsclient is more involved. It has some C99 features that my
compiler doesn't support, it uses some types that aren't present in
the TCPIPLibs, and uses the pthread library. The code is fixed up
using a sed script which then makes the code compilable. The
getaddrinfo function isn't supported in my TCPIPLibs, so I knocked
up something that kinda looks like the interface, just enough to
work.

The client code can't use pthread, so we directly invoke the threads
that would have run - they would run in series anyhow, so this works.