Allow timeouts to be specified, using the new options action.
The server is being updated to allow named options to be set, which
will allow the system to be a little more configurable. This won't
include a lot of options, I expect, but initially it will allow the
timeout to be specified, which prevents the job running for longer
than you might like - or allows it to run for longer if you have a
longer running build or test that you wish to run.