Update 'response' action handler to accept structured responses.
The 'response' server action is now defined to allow a list of a
string and structured data, which makes it simpler to return any
extra messages data to answer the request. This is used by the
'options' request, which is currently not supported by the tool,
but can be with this change.